first version of style
