--- src/App.css
@@ -1,2 +1,0 @@
-.App {
-}
