+++ src/App.scss
@@ -1,0 +1,9 @@
+@import "./shared/styles/colors";
+@import "./components/Navbar/Navbar.scss";
+
+body {
+  background: $gradient_light;
+  .header-spacer {
+    min-height: $navbar-height;
+  }
+}
