+++ src/components/StickyNoteList/StickyNote/IdeasList/IdeasList.scss
@@ -1,0 +1,24 @@
+@import "../../../../shared/styles/colors";
+
+.sticker {
+  justify-content: center;
+  margin: 10px;
+  min-width: 200px;
+  background: $oil-dark;
+  padding: 5px;
+  border-radius: 10px;
+  box-shadow: 0px 0px 10px black;
+  .sticker__title {
+    padding: 4px 40px;
+    margin: 7px 10px 0px;
+    text-align: center;
+    border-radius: 5px;
+    background: $mustard;
+  }
+  .sticker__content {
+    margin: 10px 0px;
+    // width: 100px;
+    min-height: 100px;
+    text-align: center;
+  }
+}
