--- src/components/StickyNoteList/StickyNote/IdeasList/idea/index.js
@@ -1,2 +1,0 @@
-import React from "react";
-import "./Idea.scss";
