--- src/components/StickyNoteList/StickyNote/StickerNote.scss
@@ -1,25 +1,0 @@
-@import "../../../shared/styles/colors";
-body {
-  background: $mustard;
-}
-
-.sticker {
-  margin: 20px;
-  min-width: 200px;
-  background: $oil-dark;
-  padding: 5px;
-  border-radius: 10px;
-  box-shadow: 0px 0px 10px black;
-  .sticker__title {
-    padding: 4px 40px;
-    margin: 7px 10px 0px;
-    text-align: center;
-    border-radius: 5px;
-    background: $oil_light;
-  }
-  .sticker__content {
-    margin: 10px 10px;
-    width: 100px;
-    height: 100px;
-  }
-}
