+++ src/components/StickyNoteList/StickyNoteList.scss
@@ -7,4 +7,5 @@
 .sticker-list {
   display: flex;
   flex-wrap: wrap;
+  justify-content: center;
 }
