+++ src/shared/styles/layout.scss
@@ -1,0 +1,1 @@
+$nav-margin: 20px;
